update LDi_sleepMilliseconds to use unsigned long instead of unsigned int

+++ include/ldinternal.h
@@ -113,7 +113,7 @@ struct LDClient {
 
 /* **** LDUtility **** */
 
-bool LDi_sleepMilliseconds(const unsigned int milliseconds);
+bool LDi_sleepMilliseconds(const unsigned long milliseconds);
 bool LDi_getMonotonicMilliseconds(unsigned long *const resultMilliseconds);
 bool LDi_getUnixMilliseconds(unsigned long *const resultMilliseconds);
